Files of exercises L02

+++ L01_JavaIntroduction/src/Exercise6.java
@@ -1,7 +1,7 @@
 public class Exercise6 {
     public static void main(String[] args) {
         System.out.println("a = " + (9+1)*9/2);
-        System.out.println("b = " + (long) (1000000+1)*1000000/2);
+        System.out.println("b = " + (1000000L+1)*1000000/2);
         System.out.println("c = " + (9.5*4.5 - 2.5*3) / (45.5-3.5));
         System.out.println("d = " + 4 * (1 - 1.0/3 + 1.0/5 - 1.0/7 + 1.0/9 - 1.0/11 + 1.0/13));
     }
